use number format for currency in all websites dashboard
