+++ plugins/MultiSites/tests/Integration/DashboardTest.php
@@ -71,7 +71,7 @@ class DashboardTest extends IntegrationTestCase
                 'nb_visits' => 0,
                 'nb_actions' => 0,
                 'nb_pageviews' => 0,
-                'revenue' => '$ 0',
+                'revenue' => '$0',
                 'visits_evolution' => '0%',
                 'actions_evolution' => '0%',
                 'pageviews_evolution' => '0%',
@@ -85,7 +85,7 @@ class DashboardTest extends IntegrationTestCase
                 'nb_visits' => 0,
                 'nb_actions' => 0,
                 'nb_pageviews' => 0,
-                'revenue' => '$ 0',
+                'revenue' => '$0',
                 'visits_evolution' => '0%',
                 'actions_evolution' => '0%',
                 'pageviews_evolution' => '0%',
@@ -99,7 +99,7 @@ class DashboardTest extends IntegrationTestCase
                 'nb_visits' => 0,
                 'nb_actions' => 0,
                 'nb_pageviews' => 0,
-                'revenue' => '$ 0',
+                'revenue' => '$0',
                 'visits_evolution' => '0%',
                 'actions_evolution' => '0%',
                 'pageviews_evolution' => '0%',
@@ -123,7 +123,7 @@ class DashboardTest extends IntegrationTestCase
                 'nb_visits' => 0,
                 'nb_actions' => 0,
                 'nb_pageviews' => 0,
-                'revenue' => '$ 0',
+                'revenue' => '$0',
                 'visits_evolution' => '0%',
                 'actions_evolution' => '0%',
                 'pageviews_evolution' => '0%',
