+++ plugins/SitesManager/API.php
@@ -1242,7 +1242,7 @@ class API extends \Piwik\Plugin\API
      */
     public function getCurrencyList()
     {
-        $currencies = Formatter::getCurrencyList();
+        $currencies = Site::getCurrencyList();
         return array_map(function ($a) {
             return $a[1] . " (" . $a[0] . ")";
         }, $currencies);
@@ -1255,7 +1255,7 @@ class API extends \Piwik\Plugin\API
      */
     public function getCurrencySymbols()
     {
-        $currencies = Formatter::getCurrencyList();
+        $currencies = Site::getCurrencyList();
         return array_map(function ($a) {
             return $a[0];
         }, $currencies);
